journald: Fixed potential deadlock on cancel.
The JournaldProducer function writes journald entries to a Go
channel as it reads from a systemd journal. When the provided
context.Context enters the "done" state (in this case, because
the process received a signal from the OS), the function returns
and decrements a sync.WaitGroup by one. The parent Go routine
monitors the WaitGroup and only exits once the WaitGroup is
equal to zero, indicating that all Go routines have exited.

The channel-write operation did not account for the context.Context
being marked as "done". As a result, the channel write may deadlock
if the corresponding consumer.JournaldConsumer function returns
while the "producer" is blocked trying to write to the channel.

Since the main thread only returns when the WaitGroup hits a zero
count, the program could potentially deadlock when it receives an
exit signal from the OS.

I think the overall chances of hitting this bug are low because
the channel is buffered. However, both functions are controlled
by the same context.Context. The "producer" may also sleep for
some time if no entries are available - increasing the risk of
a race between the sleep and the context being marked as done.