Only re-initialize journald reader on error
My initial reading of the SD_JOURNAL_INVALIDATE was wrong. It doesn't
only reflect a rotation, it may reflect potential records not being
sequential and many other things. AND, it is returned on success; which
is the issue here.

Instead, let's rely on error codes (negative numbers) to check if we
need to re-initialize the journal.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>