Move flush logic to producer
The producer will now log the last read log so we don't repeat. This
ensures that we still are able to output the needed event while not
flooding the logs with events.

This also makes sure to update the runtime with the current "last read"
timestamp before resetting the journal.

Finally, the wait time between errors is increased.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>