Add makefile target for testing container on live instance info
This introduces the `instance-test` Makefile target which downloads all
the relevant information from a live instance and runs a container
locally using that info.

This is useful for debugging an issue without having to run the
container on that machine.

An instance can be specified via the `LIVE_INSTANCE` variable.

It defaults to using the `core` user to scp files from a machine,
however, a user can be specified via the `INSTANCE_USER` variable.

To clean up, you can run the `clean-instance-test` Makefile target.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>