Sometimes concurrency is not the answer
This removes the concurrent producer and consumer in favor of a single
main loop that will read from the journal and process audit events.

The reasoning for this is that we need to ensure that audit logs get
processed and we flush the processed event in preference of the last
successfully written audit log. Instead of complicating the code base
with exit priorities, this removes the concurrency logic.

This also moves the logic of the multiple channels to use `errgroup`
instead. This helps us run "read loops" and receive errors from the
multiple loops as needed. This will be handy when we start processing
auditd audit logs events.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>