data: Fixed last-read journal timestamp handling.
Normally, the application saves the last-read journal event's
timestamp to a file. This file is read by the program at startup,
and its contents are used to "seek" just after the last-read
event's timestamp. This allows the program to resume where it
left off between executions.

Unfortunately, this logic was broken at some point such that it
never saves the last-read timestamp. This commit attempts to both
fix and simplify the handling of the last-read timestamp by:

- Removing the use of sync/atomic and the int pointer
- Making the current timestamp a struct field
- Making the last-read timestamp an input to the code that
  instantiates a new JournalReader object

I considered updating the "flush" (file save) logic to run
periodically. I decided that this commit already has a lot
of moving parts. Perhaps next time?