general: Cleanup handling of errgroup.Group.
This commit moves the creation of the errgroup.Group closer to
the code that uses it and renames the context.Context it returns
for readability. I also removed the errors.Is check for the
context.Canceled error and added a comment warning about why
such a check is not safe.

All of these changes are meant to make the use of the returned
context.Context more obvious. This Context is canceled when the
very first non-nil error is returned by a Go routine started by
errgroup.Group. By passing the Context into the child Go routines,
we ensure that the other running Go routines are cancelled if
an error occurs. The errgroup documentation does not make this
entirely clear.