Optimize fsnotify file name comparisons
strings in golang already have the length by default, but comparisons
will do a canonical string comparison without taking this into account.
By checking the length first, we do add an O(1) operation for every
comparison, but we optimize given the type of strings we're getting.

normally, we get a path plus the rotation number, so strings are very
similar and are mostly different because of the length (we are only
targetting the main log path without the rotation number). So, this
optimization hits all cases we're actually interested in.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>