Add `coverage` Makefile target
This runs the unit tests while also generating coverage information.
This is useful to know what codepaths we've missed with the newer test
additions.

While the goal is not to reach 100% code coverage, it is useful to know
what areas we're not actively testing as it may help us either come up
with more tests to ensure we don't hit unexpected bugs, or simply know
where we're making assumptions.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>