sessiontracker: Cleanup auditdEvent.
This commit updates the auditdEvent method so that it only
creates new audit sessions when it receives an audit event
of type AUDIT_LOGIN. Under normal circumstances, I think
the original code was "OK" - but having a more explicit
check is not only a nice safeguard, but also helps clarify
what is going on.

This commit also  adds some code comments to help clarify
what is happening in the method.