journald: Handle usernames w/ non-letters.
The existing "login" regex for public key auth events does not
match usernames that contain "-" and other characters. This commit
updates the regex to handle usernames with "-", ".", "@", and "$".

I have read that Unix account names can be very permissive (even
allowing "/"). According to the "adduser" program on Ubuntu, only
a subset of characters are permitted according to POSIX. I have
also added tests for those characters.