integration-tests: Initial take.
This commit adds a single integration test that verifies the
"happy path" for the application on Ubuntu. It can be executed
directly with "go test -tags int ./..." or via "make". The tests
assume that they are being executed in a test environment and
require root privileges in order to manipulate the OS and
interact with various daemons.

The commit also adds a vagrant VM config that mounts the source
code to "/vagrant" in the VM. Developers can use this by running:

  cd internal/integration_tests/vagrant-<whatever>
  vagrant up --provider virtualbox
  vagrant ssh

The tests can also be executed in a GitHub CI runner VM or
a custom-provisioned VM. Refer to the GitHub CI workflow and
the Vagrantfile for pre-test dependencies.

The single integration test essentially does the following:

1. Creates a new user account and a SSH CA. The tests use this
   credential to login to the newly-created user later in
   the test
2. Creates a named pipe that audito-maldito writes events to
3. Reads events from the named pipe and verifies that they are
   what we expect
4. Starts audito-maldito and points it at the named pipe
5. Connects to the local SSH server as the newly-created user
   and executes a shell pipeline
6. Confirms that the pipeline ran

Ozz requested that I structure the test code so that other
developers could easily construct tests. I attempted to do
that here... although, I am not sure if I was successful.

On a side note - the choice of terminology (integration, system,
end-to-end) is all over the place. We can change the name if
we desire.