integration-tests: Fix exec order check.
This commit fixes the verification of the SSH shell pipelne.
Shell pipelines appear to execute the second (last?) program
first. This likely has to do with redirecting the process' std
IO - which may require executing the applications inverse of
the order listed in the string passed to the shell. The test's
verification code now starts at the end of the pipeline.

This commit also makes the integration test wait until the
application's readiness check is complete *before* executing
an SSH login, thus isolating the test from things like parsing
large amounts of audit events found in existing audit logs.

Finally, this commit updates the test's verification code
to look for the SSH certificate's "certificate ID" in the
corresponding audit events.