sessiontracker: Fixed unreliable test(s).
I noticed TestSessionTracker_AuditdEvent_ExistingSession_Ended
would occasionally fail a while back. I saw it fail for the
first time in CI today. This commit definitely fixed *a* problem.
I am not sure if it is the only issue with the test, however.

Refer to code comments in the diff for details.