tests: Decrease dependence on ctx cancel.
It appears that cancelling a context may not immediately result
in it being marked as done. Looking at the Go context library,
it is not immediately obvious how that is possible - however,
I did observe it happening once in CI. [1]

Instead of relying on context cancellation, use the
recently-added "err" field in the testAuditEncoder.

1. https://buildkite.com/equinix/audito-maldito/builds/272#0186559b-d020-408e-b562-c08f04251c38
(need to view "raw" log to see the failure)

From the CI log:

  === CONT  TestSessionTracker_AuditdEvent_ExistingSession_WriteCacheErr

    sessiontracker_test.go:385:

    Error Trace: /workdir/internal/auditd/sessiontracker_test.go:385

      Error:     Target error should be in err chain:
                 expected: "context canceled"
                 in chain:
      Test:      TestSessionTracker_AuditdEvent_ExistingSession_WriteCacheErr

  --- FAIL: TestSessionTracker_AuditdEvent_ExistingSession_WriteCacheErr (0.00s)