auditd: Temporary "fix" for deadlock.
During deployment, we found that audito-maldito would deadlock
when writing a UserLogin event to the RemoteUser logins channel
(see "internal/journald/processentry.go", line 240).

We found that the go-libaudit Reassembler.Close method was
executing our reassembler call-back's methods - which results
in a write to the reassembleAuditdEvents channel. When coupled
with a "defer" statement, the Close method deadlocked because:

  - The context.Context was not marked as done
  - No Go routines are reading from reassembleAuditdEvents

The Reassembler.Close method should *really* reflect that it
calls the reassembler callback.