Consume events flushed by re-assembler `Close`
The reassembler will flush the audit events that were not fully
processed (evicted) once it's closed. Given that we now don't do this
via a channel, but instead hook up the sessiontracker directly, we can
simply assume that events will be coalesced and writen to the audit
writer.

This PR simply verifies that the closing works and flushes an event
since there's nothing blocking now.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>