Move dirreader and sessiontracker into their own packages
This makes auditd package easier to browse and the code easier to
refactor since we remove dependencies from internal fields in
structures.

The dirreader is now under the auditd package, which opens up the
possibility of moving it elsewhere if we see it being used generically.

This also centralizes the testing tools into their own package to reduce
duplication.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>