Remove one level of channels using session tracker directly
This uses the session tracker directly in the reassembler callback in
order to remove one level of channel message passing.

To do this, the sessiontracker now introduces an interface called
`Auditor` which will output audit messages as called in a non-blocking
manner in terms of threading (it blocks on IO).

Errors are still passed to a channel, but it is now a channel just
dedicated towards errors.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>