Introduce revamped health/readiness
This moves the health code into its own package. It also allows for the
enablement of a readiness endpoint (`/readyz`) if a `--readyz` flag is
passed to audito-maldito.

The health system was also refactored to take a number of components for
which we'll be able to output if they're ready or not. These components
are then listed as part of the readyz endpoint output.

In the future, the intent is to also add a livez endpoint.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>