health: Fix unit tests
The unit tests were testing values at random, but the randomness had a
logical error. That is, it may have tested 0 services for readiness,
which is not what we wanted to test.

This removes the randomness from these tests, as it's not necessary to
add entropy to test the functionality we need. This logical error is a
carry-over from the initial implementation of the readiness logic.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>