Add HTTP Server timeouts
Now that we're serving metrics and readiness through HTTP, we should be
covering timeouts for the calls to avoid client issues or attacks.

This adds a couple of command line arguments to make the aforementioned
timeouts configurable.

Also, this removes the `nolint` directive since the complexity was
addressed already. We shouldn't just add blank `nolint` directives as
we'll miss valid warnings.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>