internal/health: Fix static readiness HTTP status.
Prior to this change, the readiness HTTP endpoint would always
write a HTTP 200 status code header - even if the application
was not ready.

This happened because we indirectly call http.ResponseWriter.Write
when we stream the readiness status map to the HTTP client. This
occurs when we execute "json.NewEncoder(w).Encode(...)".

Executing http.ResponseWriter.Write results in an implicit call
to http.ResponseWriter.WriteHeader, with the default value being
a 200 status code. [1] Subsequent calls to WriteHeader are ignored
by the http library. As a result, we were always setting the HTTP
response status code header to 200. This behavior is likely meant
to reflect how HTTP should work; HTTP headers are written to the
peer before the body is written.

1. https://pkg.go.dev/net/http@go1.20.3#ResponseWriter