auditd: Make test reliable w/ TestAuditEncoder.Err,
This unit test (renamed in the commit for accuracy) was failing
intermittently due to the cancellation of the context.Context
racing with a channel write to a buffered channel (well, that
is what I think was happening anyways).

This commit updates the problematic test to use the
TestAuditEncoder.Err field, which is checked by the
Write method prior to writing the AuditEvent - thus
avoiding any races / inconsistent behavior.